Global Refactoring
A completely new architecture of the core part of the game was created, observing the design principles of SOLID. The classes responsible for the logic of heroes, monsters, towers and their interactions with each other were completely rewritten. The monster spawn system was also changed. It was implemented using abstract factory patterns and object pooling.